app/requirements.txt
psycopg2-binary==2.9.5
pandas==1.5.2
